add jsconfig file for baseUrl/paths
